I init new util: getEncodedPurchaseData.
